Fixed the 0 index bug along with tree restructure.
@freddy Turns out your intuition on the script was right. The script was
on the wrong node, so when it kept accessing unitilized name and
position. I restructured it so that movement is the root Player Node
with Kinematic Body.

  In addition, the spawning mechanism of the bullets
is influenced by the position of the parent for some reason (relative
maybe?).
So I changed back to addding the bullets as children of the map instead
of children of the combat node.