Network (#2)
* Basic server creation script

* Interim changes to Player

* Add movementspeed extern var

* Add networking to Player.gd [experimental]

* omegalul:

* Minor change

* Renamed to minus

* Added Network To Autoload

* Add main menu and net menu

* Fixed the 0 index bug along with tree restructure.

@freddy Turns out your intuition on the script was right. The script was
on the wrong node, so when it kept accessing unitilized name and
position. I restructured it so that movement is the root Player Node
with Kinematic Body.

  In addition, the spawning mechanism of the bullets
is influenced by the position of the parent for some reason (relative
maybe?).
So I changed back to addding the bullets as children of the map instead
of children of the combat node.

* Multiplayer support should work now

* Fixed error with host/join emit function

The buttons' signal function were swapped.
That 4am game jam focus :)

* Working Multiplayer!!!!!!

Now the only thing needed to do is some slight adjustements on the
remote server calls. (Disconnection on server and freeing nodes when
players leave.)

* Add that image