Merge master into Aiming branch.
Structure of Player scene tree is set to master's (so that the
root node is KinematicBody2D).

Beware when merging two scenes--the result could be
incomprehensible to Godot and result in a mysterious crash (and
also a scene file that cannot be opened in Godot).

For now, the player direction is server-authoritative, although
this might be a bad choice since you want to request a projectile
to fly in the direction that you specify as the player.